fix dispense_labware hydrated form data type
