+++ protocol-designer/src/form-types.ts
@@ -18,6 +18,8 @@ import type {
   ChangeTipOptions,
   LabwareEntity,
   PipetteEntity,
+  TrashBinEntity,
+  WasteChuteEntity,
 } from '@opentrons/step-generation'
 export type StepIdType = string
 export type StepFieldName = string
@@ -231,6 +233,25 @@ export type BlankForm = AnnotationFields & {
   id: StepIdType
 }
 
+interface LabwareEntityWithTouchTip extends LabwareEntity {
+  isTouchTipAllowed: boolean
+}
+
+interface WasteChuteEntityWithTouchTip extends WasteChuteEntity {
+  isTouchTipAllowed: boolean
+  name: 'wasteChute'
+}
+
+interface TrashBinEntityWithTouchTip extends TrashBinEntity {
+  isTouchTipAllowed: boolean
+  name: 'trashBin'
+}
+
+export type LabwareOrAdditionalEquipmentEntity =
+  | LabwareEntityWithTouchTip
+  | WasteChuteEntityWithTouchTip
+  | TrashBinEntityWithTouchTip
+
 export interface HydratedMoveLiquidFormData extends AnnotationFields {
   id: string
   stepType: 'moveLiquid'
@@ -246,7 +267,7 @@ export interface HydratedMoveLiquidFormData extends AnnotationFields {
   changeTip: ChangeTipOptions
   dispense_airGap_checkbox: boolean
   dispense_delay_checkbox: boolean
-  dispense_labware: LabwareEntity | AdditionalEquipmentEntity
+  dispense_labware: LabwareOrAdditionalEquipmentEntity
   dispense_mix_checkbox: boolean
   dispense_touchTip_checkbox: boolean
   dispense_wellOrder_first: WellOrderOption
@@ -341,7 +362,7 @@ export interface HydratedMixFormData extends AnnotationFields {
   dispense_delay_checkbox: boolean
   dropTip_location: string
   id: string
-  labware: LabwareEntity
+  labware: LabwareEntityWithTouchTip
   liquidClassesSupported: boolean
   mix_touchTip_checkbox: boolean
   mix_wellOrder_first: WellOrderOption
